requirements/base.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
python-decouple>=3.8
Pillow>=10.0
psycopg2-binary>=2.9
redis>=5.0
django-redis>=5.4
django-ratelimit>=4.1
drf-spectacular>=0.27
httpx>=0.27
pytz>=2024.1
channels>=4.0
channels-redis>=4.0
celery>=5.3
flower>=2.0
daphne>=4.0
